refactor: group 컴포넌트 관심사에 따라 한 depth 더 분리
