+++ src/components/group/detail/GroundRules.tsx
