+++ src/components/group/detail/TeamCard.tsx
