+++ src/components/group/detail/TodayGoals.tsx
