+++ src/components/group/search/EventBanner.tsx
